refactor: 요구사항에 맞게 이름 변경
HttpRequest -> HttpRequestUtils
separate어쩌구 -> parse어쩌구
