+++ src/test/java/webserver/HttpRequestUtilsTest.java
@@ -2,29 +2,30 @@ package webserver;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.util.HttpRequestUtils;
 
 import java.util.Map;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-class HttpRequestTest {
+class HttpRequestUtilsTest {
     @Test
     @DisplayName("Header 첫 번째 라인에서 요청 URL 추출 확인")
     void testSeparateUrl() {
         // 기본 URL 처리
-        String url1 = HttpRequest.separateUrl("GET /index.html HTTP/1.1");
+        String url1 = HttpRequestUtils.parseUrl("GET /index.html HTTP/1.1");
         assertThat(url1).isEqualTo("/index.html");
 
         // URL에 파라미터가 포함된 경우
-        String url2 = HttpRequest.separateUrl("GET /index.html?id=sully HTTP/1.1");
+        String url2 = HttpRequestUtils.parseUrl("GET /index.html?id=sully HTTP/1.1");
         assertThat(url2).isEqualTo("/index.html");
 
         // URL에 경로가 포함된 경우
-        String url3 = HttpRequest.separateUrl("GET /user/sully123 HTTP/1.1");
+        String url3 = HttpRequestUtils.parseUrl("GET /user/sully123 HTTP/1.1");
         assertThat(url3).isEqualTo("/user/sully123");
 
         // URL이 "/"인 경우
-        String url4 = HttpRequest.separateUrl("GET / HTTP/1.1");
+        String url4 = HttpRequestUtils.parseUrl("GET / HTTP/1.1");
         assertThat(url4).isEqualTo("/index.html");
     }
 
@@ -32,7 +33,7 @@ class HttpRequestTest {
     @DisplayName("요청 메서드 추출 확인")
     void testSeparateMethod() {
         String line = "GET /index.html HTTP/1.1";
-        String method = HttpRequest.separateMethod(line);
+        String method = HttpRequestUtils.parseMethod(line);
         assertThat(method).isEqualTo("GET");
     }
 
@@ -46,7 +47,7 @@ class HttpRequestTest {
                 "name", "%EB%B0%95%EC%9E%AC%EC%84%B1",
                 "email", "sully1234%40naver.com");
 
-        Map<String, String> paramMap = HttpRequest.separateParam(line);
+        Map<String, String> paramMap = HttpRequestUtils.parseQueryString(line);
 
         assertThat(paramMap).isEqualTo(userMap);
     }
